Added a new format for documents
